When syntax is complete.  Fixed seg fault associated with using $-s.  Will need to talk to bhupc about how to allow internal system to use $-s, but prevent external scripters from using $-s.  Still have weird segfault that occurs from an exception within import call's not being caught correctly.  Bhupc's been doing some error handling, so maybe it'll fix itself when I re-base.

+++ liboh/plugins/js/JSObjectStructs/JSVisibleStruct.cpp
@@ -26,14 +26,14 @@ v8::Handle<v8::Value> JSVisibleStruct::dist(Vector3d* distTo)
 
 
 
-JSVisibleStruct::JSVisibleStruct(JSObjectScript* parent, const SpaceObjectReference& whatsVisible, const SpaceObjectReference& toWhom, bool visibleCurrently, const Vector3d& currentPosition, const Vector3d& currentVelocity)
+JSVisibleStruct::JSVisibleStruct(JSObjectScript* parent, const SpaceObjectReference& whatsVisible, const SpaceObjectReference& toWhom, bool visibleCurrently, const Vector3d& currentPosition, const Vector3f& currentVelocity)
  : JSWatchable(),
    jsObjScript(parent),
    whatIsVisible(new SpaceObjectReference(whatsVisible)),
    visibleToWhom(new SpaceObjectReference( toWhom)),
    stillVisible(new bool(visibleCurrently)),
    mPosition(new Vector3d(currentPosition)),
-   mVelocity(new Vector3d(currentVelocity))
+   mVelocity(new Vector3f(currentVelocity))
 {
 }
 
@@ -102,7 +102,7 @@ Vector3d JSVisibleStruct::returnProxyPositionCPP()
     return *mPosition;
 }
 
-Vector3d JSVisibleStruct::returnProxyVelocityCPP()
+Vector3f JSVisibleStruct::returnProxyVelocityCPP()
 {
     return *mVelocity;
 }
